Move EventDispatcher to 'core.hooks' package
